+++ NEWS.md
@@ -1,3 +1,5 @@
+# tutorial.helpers (development version)
+
 # tutorial.helpers 0.2.7
 
 * Get test cases for `format_tutorial()` working again.