+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Add show_file().
 
+* Add process_submissions().
+
 # tutorial.helpers 0.2.7
 
 * Get test cases for `format_tutorial()` working again.