+++ NEWS.md
@@ -1,4 +1,4 @@
-# tutorial.helpers (development version)
+# tutorial.helpers 0.3.0
 
 * Add show_file().